[Hd] Unpickable objects should act as occluders and not be ignored altogether during picking
For picking through Hd, we currently exclude the unpickable prims from the ID render. This results in occlusion not being honored during the picking process.

To accommodate occlusion of unpickable prims, HdxIntersector executes two tasks corresponding to the two render passes.
The first pass renders the unpickable prims into the depth buffer alone. It uses a collection with the root and exclude paths swapped and masks off all color writes as part of its render pass state, which allows us to bind the same fragment shader in both passes.
The second pass is as before, wherein the pickables are rendered, thereby populating the ID render targets for fragments that pass the depth test.

Changes:
- Extended HdxIntersector to do a depth only pass with the unpickable prims.
- Added ColorMask to HdRenderPassState with the necessary boiler plate.

Testing:
Added testHdxUnpickablesAsOccluders, which draws two cubes behind each other. When the front cube is unpickable, we no longer "pick through" to the second cube.

(Internal change: 1803468)
(Internal change: 1803582)