Make HdStCommandBuffer more resilient to unset geometric shaders on draw items.
In order to get the command buffer to rebuild batches when the geometric shaders are unset, check the geometric shader in HdDrawBatch::Validate() and Rebuild().

In _RebuildDrawBatches, skip batching items with no geometric shader.

In all of these places, keep a TF_VERIFY to alert us, but if we skip drawing items with no geometric shader we avoid the null pointer exception while drawing at least.

(Internal change: 1804542)