Initial implementation of spec copying in sdf
This change adds a general-purpose function for copying
scene description in sdf. This function uses callbacks
to determine which fields and child objects to copy and
any modifications to the data that are needed during
the copying process.

An upcoming change will add a simple API on top of
this function that is aimed at basic use cases.

(Internal change: 1805575)