Add simple API for copying specs to sdf
This change adds an overload of SdfCopySpec
that does not require any callbacks and does a
simple recursive copy of the specified spec,
clearing out any pre-existing data. Relationship
targets and attribute connections that target
objects beneath the source spec will also be
remapped to the destination spec.

(Internal change: 1805758)