Added API in UsdUtils to compute and author a collection that includes a specified set of subtree paths. A vectorized version of the API is provided for convenience, for computing a set of collections in parallel and authoring them serially.
For each collection to be created, the algorithm computes an optimal set of
paths to include and paths to exclude based on the values of certain parameters.
For example, if all (or most) descendants of an ancestor path are included in the
collection, the ancestor itself will be included in the collection (while excluding the
bits that don't belong to it). Please refer to the documentation (in usdUtils/authoring.h)
for details on the parameters that control the algorithm and other details.

(Internal change: 1805912)
(Internal change: 1806036)