Update pxrUsdIn to import new style of collections encoded using UsdCollectionAPI.
If the collection is "explicitOnly", it gets baked into a flat list of member paths. If it's
'expandPrims', then we construct a CEL expression that's equivalent to the result of
includes and excludes paths as authored in USD.

Old-style collections (encoded using UsdGeomCollectionAPI) are only imported
if the env-setting USD_KATANA_IMPORT_OLD_STYLE_COLLECTIONS is set to true
(which is the current default).

(Internal change: 1805914)