Revert authoring of "prepend" list operations by default
This behavior was enabled by default in change 1802249.
However, testing and user feedback revealed this led to
confusing situations. For instance, to author a series
of references in strong-to-weak order, users would need
to call the authoring API in USD in weak-to-strong order,
since the new behavior would always add elements to the
front of the "prepend" list. This is unintuitive and
would require users to update any preexisting code to
maintain the desired strength ordering.

While work is being done to address these issues, we're
reverting to the old behavior of authoring "add" list
operations by default so that users won't be affected.

(Internal change: 1807008)