Added support for CEL expressions in baking light linking.
This refactors how we represent the links during baking.  If the
links are dynamic (i.e. at least one link isn't a simple location
but instead some more complex CEL expression) then we bake out the
raw patterns to custom blind data.  Otherwise we bake out the list
of locations as before.  We could also expand the CEL into the
locations that currently match for the benefit of naive clients
that don't know about the blind data.  But this would have a big
impact on baking performance and would yield brittle results, so
the relevant expansion code has a XXX comment and returns an
empty list.

PxrUsdIn now preferentially uses the blind data if it exists.

(Internal change: 1809663)