[Usdview] Add stage data model and unit tests.
The current Usd.Stage is now stored in a StageDataModel. Replacing the stage emits a signal which can be handled by anything with access to the data model, including other parts of the data model which need access to the current stage.

The "close stage" timer was moved into the StageDataModel because now it deletes the final reference to old stages.

(Internal change: 1811008)
(Internal change: 1811029)
(Internal change: 1811240)