[usdMaya] When exporting reference assemblies, make sure there is always an xformOpOrder -- identity transforms should have a value block to force them to be identity
Say that you have a model ModelWithScale that has a scale xformOp at its root.
When you bring it into Maya, the xformOps at the root are replaced with the xformOps that
come from the Maya reference assembly node. This means that, when the reference
assembly has an identity transform, it is not enough to write out *no* xformOps to the
prim. We must also *block* the xformOpOrder so that any existing xformOps get ignored.

This really only affects the identity case, because if the reference assembly is moved
around, then we properly override the existing xformOpOrder with our new
xformOpOrder.

(Internal change: 1811396)