[usdMaya] fixes for import and export of camera apertures/widths
This change corrects the handling of the aperture for orthographic cameras.
Orthographic cameras in Maya use the 'orthographicWidth' attribute to
determine the size of the image, and the aspect ratio, lens squeeze ratio,
film aperture and film offset attributes have no effect. This means that on
export, we need to translate Maya's orthographicWidth to USD's
horizontalAperture and verticalAperture, and we do not need to author
horizontalApertureOffset and verticalApertureOffset. Note that it currently
doesn't seem possible to create a non-square orthographic camera in Maya.
We do a similar thing in reverse on import, though we also bring in all of
the other perspective-only attributes as well, just to try to preserve the
data.

This change also fixes an issue where lens squeeze ratio was being
incorrectly applied to the vertical film aperture. Lens squeeze ratio
should only apply to the horizontal dimension for anamorphic cameras.

Finally, film offsets need to be converted from inches to millimeters on
on export.

(Internal change: 1811401)