[Usdview] Fix 'shut_down_hydra' and 'close_stage' timing
Performance tests after change #1811008 indicated that Hydra shutdown times increased significantly and USD Stage close times dropped to zero. Before this change, the last reference to the renderer was deleted then the last reference to the stage was deleted (effectively closing both), but the order was switched after the change. The problem with this is the renderer stores its own reference to the stage so both objects are now deleted together rather than separately. This causes the "close_stage" time to drop to zero because we are not actually closing it when we expect to. Instead, the time from "close_stage" was added to "shut_down_hydra," causing an apparent increase.

The fix for this is to switch back to closing the renderer before closing the stage.

(Internal change: 1811410)