Allow '?' in layer identifiers
This change lifts the restriction on using '?' in layer
identifiers. This restriction had been in place because
that character was reserved as a delimiter for
encoding SdfLayer::FileFormatArguments in an
identifier. Instead, these arguments are now delimited
with the string ":SDF_FORMAT_ARGS:".

The lex/yacc-based layer identifier parser was also
removed, since it was not providing much benefit
over a simple hand-rolled parsing function.

Fixes #289

(Internal change: 1811743)