[Usdview] Created a data model to encapsulate selection.
Previously, selection state was stored in multiple Qt widgets (mainly primView and propertyView), as well as in scattered properties in appController and stageView. All of this data was centralized into the selection data model which has a clean API which includes signals for when the selection is modified.

The selection data model has 3 separate parts:
1. Prim selection. A set of UsdPrims. If a PointInstancer is selected, individual instance selections are also recorded in the prim selection.
2. Property selection. A set of USD properties (UsdAttributes and UsdRelationships). This selection is independent of the prim selection: if some of a prim's properties are selected, the prim itself need not be. Also included are a property's targets/connections.
3. Computed property selection. A set of Usdview-specific computer properties. Usdview computed properties are selected separately from Usd properties so they are not confused with each other. The computed property selection is also independent of the prim selection, but does not allow selection of targets/connections.

All parts of the selection data model emit signals when they are modified in any way. The prim selection includes added and removed prims with its signal to allow for more efficient handling of large sets of prims.

(Internal change: 1815988)