[usdMaya] better handling of shader color and transparency/presence during export
This change cleans up the handling of shader color and transparency/presence
when trying to determine the displayColor and displayOpacity of an exported
Gprim. If the shader supports the MFnLambertShader function set, that is
used to access the color and transparency as it has been previously.
Otherwise we try to access attributes "outColor" or "color", in that order,
to determine the shader color, and "outTransparency", "transparency", and
"presence", in that order, to determine the shader transparency. "presence"
in particular was added to support RenderMan for Maya shaders.

(Internal change: 1816128)