[Usd] Add isMultipleApply + isPrivateApply to usdGenSchema.
'isPrivateApply' indicates that a consumer wants to layer additional
behavior on top of standard Apply generated by USD for API schemas.

'isMultipleApply' indicates that the generated Apply method for API
schemas should take a third parameter 'name', which will be appended
to the schema type name when authoring to apiSchemas.

Currently only CollectionAPI can take advantage of this, but there
will be more multiple apply, as well as private apply schemas in the future.

(Internal change: 1816325)