[usdMaya] remove usages of UsdRi schemas in favor of UsdShade
This change also addresses a few issues with "pxrRis" shadingMode export
and import.

Note that for "pxrRis" shading mode, we continue to author the source Bxdf
using the UsdRiMaterialAPI during export for backwards compatibility with
consumers that may not yet have been updated to use the material's "surface"
output. We can do this without actually authoring UsdRi schema prims.
Similarly, during import we fall back on reading the source Bxdf if the
"surface" output of the material is not authored.

(Internal change: 1816474)