[Usdview] Added instance selection by instance ID when a PointInstancer has authored IDs.
Since the renderer uses instance indices (not IDs) for mouse selection and selection highlighting, logic was added in AppController.onPrimSelected and StageView._updateSelection to convert to and from IDs when they are authored. A new test case for this logic was added to testUsdviewSelectionHighlighting.

Before this change, instance indices are always used to identify selected instances. Since this is no longer the case, the names of several methods and variables (especially in selectionDataModel.py) were modified to use 'instance' instead of 'instanceIndex' to avoid confusion.

(Internal change: 1816609)