Usd: ListPosition API tweaks.
The goal of UsdListPosition is to provide both sensible default
behavior when adding list-edited items as well as an extra level
of control where needed.  Unfortunately, attempting to roll out
pipeline changes to use the current set of UsdListPosition values
revealed that they weren't ideal, so this change brings some
tweaks.

Specifically, in the case of doing a for loop to add items A B C,
we'd like the result to look like A B C.  With the current default
behavior (UsdListPositionFront), the result is C B A, since each
item is prepended to the prependedItems list.  You can get what
you want here if you reverse the order of the for loop, but
doing that and explaining why it's necessary made us realize this
was less than ideal (and it's easy to forget).	So, instead
of UsdListPositionFront, we think a better default would be to
append to the prepend list, with a new value:

    UsdListPositionAppendToPrependList

See common.h for full details.
For consistency, we also provide corresponding variations:

    UsdListPositionPrependToPrependList
    UsdListPositionAppendToAppendList
    UsdListPositionPrependToAppendList

While verbose, these names faithfully reflect the underlying
mechanism, so we think this is the right way to make the
choices comprehensible for cases where a choice is necessary,
and also provides a good default for the majority of cases.
The backing logic lives in Usd_InsertListItem() in valueUtils.

These new values also allow us to roll back some of the tweaks
made in change 1802249, such as doing reverse for loops().

As a change in behavior, this change also makes AddVariantSet(),
if called with an existing variantSet name, move the variant
set to the requested list position.  This is consistent with
how other forms of list editing API work in Usd.

(Internal change: 1816941)