Basically, there was a lot of glsl error being printed in the console when opening a usd shot file. Many errors caused by a redeclaration of HD_HAS_input.
Our UsdImagingHydraMaterialAdapter::GetMaterialParams was using attribute.GetName() which would return the name of the parameter with "inputs:", which was expected but not really what we wanted. After this change, it will now use GetBaseName() which will return the actual name without input prefix (if it is there, so it is backwards compatible).

(Internal change: 1817545)