[Usdview] Move properties from appController to a new view settings data model.
Most of these properties were view properties which influence the behavior of StageView. The exception is the `playing` property which fits better in rootDataModel because this is where this is where the current frame is stored. Several properties still had value checks in general appController code rather than in their @property methods so this was corrected.

Also add unit tests for value-checked and fully-procedural properties in viewSettingsDataModel.py. settings2.py was modified to allow for StateSources to exist without a Settings2 object so ViewSettingsDataModel could be tested.

(Internal change: 1817546)