UsdMaterial: When reading a whole materialGroup from an asset, respect the recommended katanaPath rather than the prim name
the problem was in the fact that LooksGroup op only gets triggered by its parent: in our case, a model. If we set a subroot that is not the model itself, that step is bypassed and the materials are resolved individually. If you resolve a location individually you don't have a way to set the name/path of the location, only its attributes. The bootstrap sets the correct op to load all the materials at the group level.
that is pxrUsdInLooksGroupOp instead of just pxrUsdIn

(Internal change: 1817887)