[Hd] Fixed triangle pair generated for quads and smooth normals normalization issues. The geometry shader logic that generates a pair of triangles from a quad used an incorrect split, wherein the center vertex that is generated via quadrangulation wasn't part of both triangles. This resulted in differences in lighting with the legacy imager due to normal interpolation differences.
Related to this issue is the way primvars, in particular, smooth normals, are quadrangulated.
Hydra computes (via the CPU or GPU) the smooth normals at each vertex in the coarse topology. These normals are normalized (as expected).
However, when quadrangulation follows this step, the normals computed for the edge and center vertices that are inserted aren't normalized.
In Hydra, quadrangulation may be done either on the CPU (templatized code) or GPU. In either case, we don't pass an additional 'needsNormalization' argument, and thus end up with un-normalized normals for edge and center vertices. As a catch all fix, this CL normalizes non-zero normals in GetVertexNormal in meshNormals.glslfx.

Changes:
- Updated mesh.glslfx to generate the correct triangle pair for a quad.
- Updated meshNormal.glslfx to return the normalized vertex normal for smooth normals, and compute the correct flat normal for quads.
- Updated meshWire.glslfx for the triangle pair generated from a quad.
- Fixed a typo in the compare for testHdStBasicDrawing_MDI (hdSt/testSpecs.xml)

Testing:
Several test baselines have been updated due to differences stemming from:
(a) rasterization of quad-based topology: Minor differences in fragment coverage due to the triangle pair emitted
(b) normals: Most differences are due to fragment normal differences due to the triangle pair emitted. Normalization at the VS might have played a part too for quadrangulated geometry.

(Internal change: 1817924)