Muted layers cannot be un-muted
This occurs when a layer is muted before it is opened
by calling SdfLayer::AddToMutedLayers.
The problem is that each layer keeps track of its
file modification time to avoid reloading layers.
Since an unmuted layer's file modification time was
not read, it was mis-interpreted as a newly
created file that doesn't have a file modification time.

The fix is to grab file modification time even if
the layer is muted.

(Internal change: 1818675)
(Internal change: 1818751)