Fix bug with population masks beneath instances
Creating a stage with a population mask that included
prims beneath instances would cause masters for those
instances to not be created. This was because Pcp did
not invoke the name children predicate on ancestors
of the prims in the mask during initial stage
population, so instances would never be registered
with the instancing cache.

Usd now applies the population mask in the name children
predicate instead of asking Pcp to compose the masked
paths directly. This ensures that Pcp composes the
ancestors of the masked paths and applies the predicate
for each of those indexes.

Fixes #312

(Internal change: 1818809)