Fix for relocating prims with the same source name
Fixes a bug where a relocated child prim would be
ignored if another prim with the same source name had
already been relocated in a separate composition arc.

This was because the source prim name was being
marked as prohibited the first time it was seen and
ignored when processing all subsequent nodes. To
fix this, Pcp now defers removing prohibited names
until the very end of the child names computation.

In addition, Pcp now also fixes up relocated subtrees
during prim indexing to ensure that opinions that
had been relocated deeper in the composition graph
don't contribute to prims that had been introduced
and relocated further up the composition graph.

(Internal change: 1818835)