[usdview] add a "Save Flattened As" menu item
Adds a new menu item to usdview, "Save Flattened As", which does as it says - saves a flattened version of the scene, as opposed to just a set of overrides.  This can be handy when doing debugging, or providing simplified versions of scenes to send to others, etc.  Particularly handy when used in combination with deactivating of prims, to get a "trimmed down" version of a scene.

(github pull request #346)

(Internal change: 1818896)