[HdSt] Fix failed verify of drawItem->GetMaterialShader().
A refactoring change mid-december made material shader a per-draw-item field, but the rprim classes were never updated for this change.  Material shader now needs to be set in the same pattern as geometric shader: when it's set for one draw item, it needs to be set for all, and NewRepr needs to trigger a refresh as well.

(Internal change: 1819199)