Fix for updating property types in usdGenSchema
usdGenSchema would not generate the correct code when
changing an attribute to a relationship or vice versa
in a schema. This was because usdGenSchema relied on
Usd to determine a property's type, and that relied
on the installed generatedSchema.usda, not the modified
schema in the source tree. usdGenSchema now uses Sdf
to determine a property's type to avoid this problem.

(Internal change: 1819301)