[pxrUsdMayaGL/usdMaya] initial refactoring towards Hydra collection usage revamp
The current Maya/Hydra integration is based on an early iteration of the
Hydra API in which collection names had particular meaning. This is why all
of the geometry is grouped under a single collection name. This leads to
needing to rebuild the collection at every draw which can be very expensive
and in most cases does lots of unnecessary work.

Going forward, Maya shapes that want to image using Hydra should manage
their own collections so that they can more finely control when that
collection becomes invalid and needs to be rebuilt. This change starts to
loosen the coupling between the batch renderer and individual shape
renderers. The QueueShapeForDraw() methods moved from the shape renderer to
the batch renderer. In subsequent changes, ownership of the shape renderers
may move from the batch renderer to the draw overrides/shapeUIs that draw
the Maya shapes.

The UsdMayaProxyDrawOverride::getShape() method was moved to
UsdMayaProxyShape::GetShapeAtDagPath() since it is not drawing-specific,
and it may become a common pattern as the batch renderer is extended to
support imaging other Maya shapes using Hydra.

This change also includes a bunch of cosmetic clean up. It was discovered
during that clean up that the proxy shape's getInternalValueInContex()
override was mistakenly calling the base class' set...() method for plugs
other than fastPlayback, so that has been corrected.

(Internal change: 1819801)