[pxrUsdMayaGL] fix retrieval of shape node in UsdMayaProxyShapeUI::select()
It turns out that the new UsdMayaProxyShape::GetShapeAtDagPath() method
cannot be used from the legacy viewport proxy shape UI, since the dag path
to the shape node is not readily accessible. Using selectPath() on the
MSelectInfo that's passed in yields the path to the parent of the shape
node rather than the shape node itself. As a result, we get the shape node
by just casting the MPxSurfaceShape as had been done previously.

(Internal change: 1819940)