FrustumCullIsVisible() incorrectly culls geometry
When tiny prim culling is enabled, it uses an heuristic to determine if a prim needs to be culled. However, the results are not always correct and we have cases where prims are not being rendered but they should have been drawn. For now, we will disabling tiny frustum culling by default. If you still want to use it, you can change the environment variable, keep reading for more details.

To enable/disable tiny prim culling there are two switches, the first one is via the env variable HD_ENABLE_TINY_PRIM_CULLING (false by default) which is a high level switch. If this env variable is enabled then we also provide a runtime switch in the shape of a TF_Debug with HD_DISABLE_TINY_PRIM_CULLING. When tiny prim culling is enabled, if you want to configure it you can use the DrawingRange provided by the Hd RenderPassState base class.

Fixes #314

(Internal change: 1820196)