Remove use of an ArResolverContext from all of usdview's stage-opening calls.
ArResolverContexts are generally necessary only when a tool is
opening multiple stages, and desires for each stage to resolve the
same asset paths differently.  The context that usdview was
previously creating was therefore unnecessary, inconsistent with the
initialization behavior of other tools (like usdcat), and due to a
quirk of Pixar's internal resolver, was imposing an overly
restrictive behavior - that the path used to construct an asset-based
context be pre-resolved.

Now, we allow the UsdStage to construct a default context.

(Internal change: 1820487)