[pxrUsdMayaGL] more refactoring to disentangle shape renderers from the batch renderer
This change further separates the implementation details of the
(unfortunately named) ShapeRenderer from the batch renderer. ShapeRenderers
will encapsulate the state of the shape they provide drawing services for,
they will updates this state as the shape changes, and they can provide
this state to the batch renderer when needed for drawing. With this, we
should be just about ready to split off ShapeRenderer into its own
files/class.

(Internal change: 1820784)