[pxrUsdMayaGL] extract ShapeRenderer from batchRender.h/.cpp as PxrMayaHdShapeAdapter
This change breaks out the ShapeRenderer class from being a nested class in
UsdMayaGLBatchRenderer to its own standalone class. It has also been
renamed to PxrMayaHdShapeAdapter to better indicate the function it serves.
It does not do any actual rendering, but instead provides a means of
translating from Maya scene graph data and viewport state into the
appropriate data for Hydra. PxrMayaHdShapeAdapter may ultimately become an
abstract class with a USD-specific derived class that supports Hydra
drawing of USD proxy shapes. This would allow other shape adapter classes
to be written for shapes that do not have anything to do with USD but for
which it is still desirable for them to be imaged by Hydra.

(Internal change: 1820948)