[pxrUsdMayaGL] move UsdImagingDelegate population from batch renderer to shape adapter
Since not all of the Maya shapes that the batch renderer provides Hydra
drawing for will necessarily be USD-based, the batch renderer should not be
responsible for populating the UsdImagingDelegates for USD-based shapes.
Instead, that job moves to the shape adapter, and the population takes
place during the preparation/sync stage rather than at draw time.

(Internal change: 1821058)