Fix relocation issues with PXR_INSTALL_LOCATION
When PXR_INSTALL_LOCATION was not specified at CMake
time, the USD build would add the absolute path to the
install directory to the plugin search path. This caused
problems with build relocatability: if the build was
moved around, both the old and new locations would be
searched for plugins, which could lead to odd runtime
errors.

This fix avoids defining PXR_INSTALL_LOCATION if it
hasn't been specified by the user.

Fixes #363

(Internal change: 1822155)
(Internal change: 1822450)