[pxrUsdMayaGL] first pass of moving Rprim collection management to shape adapters
This change moves the batch renderer away from managing its own single
"geometry" Rprim collection in favor of having each shape adapter manage a
collection for the shape it handles. The shape adapter will have better
insight into when the shape changes and therefore when the collection needs
to be dirtied.

(Internal change: 1822626)