Restore authoring to explicit list in various Add APIs
Change 1816941 inadvertently changed the behavior of
the Add... functions in Usd that accept a UsdListPosition
so that any pre-existing explicit list would be overwritten
by either a prepend or append list. Aside from being
surprising behavior for users, this allowed opinions from
weaker sites to contribute to the composed result, which
is likely not what the user wanted if the original list was
an explicit one.

This change restores the previous behavior so that calling
Add... on a list that is in explicit mode will modify the
explicit list. In these cases, only the position specified in
the UsdListPosition argument is respected; the append
or prepend list itself is ignored.

(Internal change: 1823426)