[pxrUsdMayaGL] remove use of boost smart pointers and noncopyable
This change replaces the use of boost smart pointers with C++11 smart
pointers. It also changes UsdMayaGLBatchRenderer to use TfSingleton instead
of boost::noncopyable, and brings the batch renderer more in line with the
singleton pattern used elsewhere in pxr.

With this change, the only remaining use of boost in pxrUsdMayaGL is
boost::hash_combine().

(Internal change: 1826403)