[pxrUsdMayaGL] remove delegate and collection params from PxrMayaHdRenderParams
The batch renderer buckets shapes for rendering based on the parameters
used to render them in an attempt to minimize Hydra/GL state changes.
However, some of the parameters that had been living in
PxrMayaHdRenderParams were in fact managed by either the shape's
HdSceneDelegate (timeCode and refineLevel), or by the HdRprimCollection
(collection name, repr name, and render tags). This change removes those
params from PxrMayaHdRenderParams so that changes to them do not cause
the shape to change buckets.

(Internal change: 1826745)