[Hd] Implement Multi-threaded Gather.
Initial attempts to implement the Predicate phase of the Prim Gather algorithm failed to show any performance benefit.  Investigation revealed a few gotcha with the original attempt.

Firstly, the TBB overhead was significant.  The Work library, uses tbb's default partitioner and a grain size of 1.  It was turning out that this partitioner was partitioning up the work completely, so it was 1 prim per task, giving a significant overhead for a large number of prims.  This is solved in this checkin by using TBB's parallel_for directly, using tunable values for controlling the overheads.

The second issue was the path translation cache in the UsdImagingDelegate.  This contained a reader-writer lock and threads were stalling on this lock. Which is fixed by a previous checkin.

Changes:
- _resultRanges in now a _ConcurrentRangeArray which is based on tbb::enumerable_thread_specific.
- _WriteResults() function is now a templated function based on iterators (so it works with Concurrent and normal ranges)
- Added _DoPredicateTestOnRange function to process the outer loop over each range.
- If the number of ranges is larger than the tunable threshold call _DoPredicateTestOnRange in parrellel using a work area.
- If the number of prims in a range is larger than the tunable threshold call _DoPredicateTest in parrellel using a tbb::parallel_for.

(Internal change: 1826930)