[pxrUsdMayaGL] move shape adapter ownership from batch renderer to UsdMayaProxyShapeUI/UsdMayaProxyDrawOverride
The UsdMayaProxyShapeUI and UsdMayaProxyDrawOverride classes have better
insight into the life cycle of individual proxy shapes, so they now take
ownership of the PxrMayaHdShapeAdapter for their proxy shape. They are also
responsible for adding and removing their shape adapter to the batch
renderer, and updating the shape adapter when their proxy shape changes.

This change also re-orders the declarations in UsdMayaProxyDrawOverride to
match the order in MPxDrawOverride.

(Internal change: 1827290)