[pxrUsdMayaGL] implement Viewport 2.0 native selection API for UsdMayaProxyDrawOverride
This change implements the wantUserSelection() and userSelect() methods of
the MPxDrawOverride API for UsdMayaProxyDrawOverride in Maya 2018 and later
so that selection will work natively in Viewport 2.0. This should remove
the requirement that the MAYA_VP2_USE_VP1_SELECTION environment variable be
enabled to correctly support selection of USD proxy shape nodes.

(Internal change: 1827643)