Remove composition error when prim specified by inherits or specializes arc does not exist.
Typically, a prim in an asset inherits from (or specializes)
a class to allow users to override that class downstream. Users
don't generally want to put any opinions for the class in the
asset itself. However, Pcp required at least one such opinion to
exist and would issue a composition error otherwise. This forced
users to create empty prims in assets just to satisfy this
requirement. Removing this error allows users to get rid of
this unwanted cruft and makes these arcs simpler to use.

This change also renames the BasicClassOverride test case
to BasicInherits so that it can contain more than just a single
example of inherits behavior.

(Internal change: 1828021)