[pxrUsdMayaGL] Add Hydra selection tracking, and render setup and selection tasks per render task
This change updates the Maya Hydra scene delegate to return a vector of
render tasks for a particular render params hash. Instead of just a single
render task, this vector will include a render setup task, a render task,
and a selection task, in that order. Combining the render setup and render
tasks is considered compatibility API, so we separate them into two tasks.

In the batch renderer, we also use ResolveUnique() instead of ResolveAll()
on the intersector results, since we are only interested in the single
nearest hit for each intersected object.

Note that while this change makes Hydra aware of the current selection,
we're not really taking advantage of that yet, but subsequent changes may
be able to leverage this.

(Internal change: 1828191)
(Internal change: 1828382)