HdGeomSubset: Change processing.
UsdGeomSubset is an interesting case because it does not
directly contribute prims to the render index.  Instead,
it affects the topology of its parent prim.  The approach
used here is to avoid adding an adapter for subsets,
and instead re-use the Mesh adapter for subset paths.
This requires the Mesh adapter to handle the case of
being invoked on behalf of a subset.

Currently, it also requires special delegate handling in
the case of adding a new subset.  It's possible that
this would be cleaner with a new virtual on adapters
that the delegate would invoke, giving them a chance
to dirty/update other parts of the scene.

A few smaller things:
- geom subset indices are now VtIntArray
- const-ify things

(Internal change: 1828495)