[Hdx] Support single render target in resolve task.
We have an internal task-specific render graph that does a single offscreen render pass to a render target.

For conveniance and to reuse the HdxDrawTargetResolveTask with this rendergraph (rather than duplicate the task), the task was update such that drawTargetRenderPasses task context entry could be either an array of unique pointers or a single HdxDrawTargetRenderPass pointer.

Changes:
- Added some typedefs for the stl types used.
- Switched on type of the drawTargetRenderPasses to build the vector of draw targets to resolve.

(Internal change: 1828892)