[HdSt] Fixed patch param refinement for Loop meshes
Hydra builds upon OpenSubdiv's patch parameterization
in order to correctly interpolate primvar data for refined
meshes. The refined patch parameterization is also
important for correctly identifying the relationship
between refined faces and authored topological
faces, edges, and points.

We found a bug in the way that OpenSubdiv computes
the patch parameterization for Loop meshes. This bug
is fixed in OpenSubdiv 3.3.1.

Making use of this fix requires implementing a method
to interpolate barycentric locations on refined triangles
in the mesh geometric shader.

Finally, face-varying primvar values need to be triangulated
for Loop meshes for both refined and unrefined topologies.

(Internal change: 1829428)