o Added a UsdAPISchemaBase class, which now forms the base class for all API schema classes. Moved the implementation of Apply() into a private member function of UsdAPISchemaBase that returns a UsdPrm upon success and added two protected, templated helper methods _ApplyAPISchema() and _MultipleApplyAPISchema() for doing the single & multiple-apply operations respectively.
o The Apply() methods now take a UsdPrim instead of taking a stage and a path,
since they require a valid prim to operate on anyway.

o Improved the documentation for the Apply methods.

o Immortalized the 'primName' tokens used in the apiSchemas metadata  by
adding them to a static tokens list in the auto-generated cpp files of API schemas.

o Updated class templates and usdGenSchema to work with the above changes.

o Removed unused private constructor in UsdCollectionAPI. Made the constructor
that takes a prim and a collection name public and added better documentation
it. Removed the GetCollection method that takes the same arguments as the now
public constructor, since clients can invoke the constructor instead.

o Templated code in UsdAPISchemaBase::_MultipleApplyAPISchema() now requires that
all multiple-apply API schemas implement a public constructor that takes a prim
and a 'name' token.

(Internal change: 1829752)
(Internal change: 1829753)
(Internal change: 1830204)
(Internal change: 1830243)