Drop unused prim indexes when changing stage population mask
Mark the pseudo-root as significantly changed to ensure that
all prim indexes are dropped before recomposing with the
new population mask applied. This makes setting the mask
slower but it helps minimize the amount of resources used
by the stage, which is one of the key purposes of the mask.

(Internal change: 1830131)