[pxrUsdMayaGL] refactor shape adapter sync process so it can be invoked directly
This change adds a GetRenderIndex() method to the batch renderer that
allows clients to get the render index they need to construct their
delegates. With this, we no longer need to make sure that the right
properties on the shape adapter are populated on the shape adapter *before*
we add it to the batch renderer. This also ensures that we rebuild the
shape adapter's delegate and update its delegate ID if any of the data it
uses to populate itself changes.

This also makes PxrMayaHdShapeAdapter::_GetWireframeColor() a static method.
This method may/will likely prove helpful for other use cases in the future,
so it may eventually be moved out of the shape adapter class.

(Internal change: 1830170)