[HdSt] Fix for modifying texture bindings that change ptex state.
In Hydra, there is a binding from the Rprim (mesh) to an Sprim (material) to a Bprim (texture).  The Mesh implementation in stream follows these bindings to see if a bound texture is a ptex and if so, enables quadrangulation.

This state change impacts topology and primvars on the Rprim.  Therefore, to handle changing the material binding on the mesh, the dirty bit propagation mechanism is used to also dirty the topology and primvar state.

Changing the material binding is handled using the dirty bit propagation mechanism.

Changing the texture binding is more tricky, as the material does not know what meshes are bound to it.  This check-in, solves this with a very big hammer, but dirtying the material binding on all Rprims.  While it is recognized
this is not the best solution, it is more correct than the code is now.  A task has been added to address this in the future.

Changes:
- If the material id is dirty on a mesh, propagate the dirty bit to primvars and topology.
- If the params change on a material, set the material id to dirty on all rprims.

(Internal change: 1830254)